Hacked up gollum so that it works better for me as a personal wiki.
I hacked up gollum so that I stayed in livepreview mode from page to page. When I click on a link in the livepreview, it takes me to a page in livepreview. That is, I don't have to go to the read-only view of the page and then click on the "edit" button. It takes me directly to the livepreview "edit" mode.

I also used r.js to optimize the livepreview AMD module so that it is just a single js file called livepreview-optimized.js.

This is really just for me to use. Just some simple hacks so that I can use this for my personal wiki.